// src/app/interfaces/ObjectAttribute.ts
export interface ObjectAttribute{
    
    att_CODE: string;
    att_ID: number;
    att_NAME: string;
    att_TYPE: string;
    bo_ID: number;
    choice_LIST: number;
    dep_ATT_ID: number;
    dependent: number;
    description: string;
    do_ID: number;
    label: string;
    last_MODIFIED_BY: string;
    lookup_ID: number;
    mandatory: number;
  
}